feat(core): refactor Pricing component from Product Card to be reusable
